rename files and links in container
